Refactor large backend files into focused submodules
Split three oversized files into smaller, single-responsibility modules:

- utils/state.py (930 lines) → state_core (variables), state_votes,
  state_clicks, state_processors, state_diversity, state_media_lookup.
  Main state.py becomes a re-export facade.

- routes/detectors.py (1564 lines) → detectors_crud (CRUD + pregen),
  detectors_training (vote/label training + multi-find),
  detectors_scoring (scoring + extractor/localizer execution).
  Main detectors.py becomes a facade with nested blueprints.

- routes/datasets.py (1101 lines) → datasets_loading (background
  threading, staging, origin management), datasets_ui (dashboard,
  demo list). Main datasets.py retains route handlers + facade.

All existing import paths preserved via re-exports. Tests updated
to use state_core for direct scalar attribute access.
